Les demi-pentes, et trois defauts qu'elles ont reveilles
Deux cases pour monter d'une : un pixel toutes les deux colonnes. Ce
n'est pas un arrondi — un demi est exact, et deux cases voisines se
raccordent toujours au pixel pres. Six formes en tout : deux directions
a quarante-cinq degres, quatre demi-cases.

Elles S'EXCLUENT, et l'editeur ne les propose donc plus en cases a
cocher : une case n'a qu'une surface, et « solide et montant a droite »
se peignait sans que rien ne dise que l'enregistrement en perdrait la
moitie. Seul « blessante » se coche par-dessus — une rampe herissee de
pointes existe.

Ecrire les demi-pentes a mis au jour trois defauts qui dormaient, dont
aucun n'appartient aux demi-pentes :

1. UNE PENTE MONTANT A GAUCHE SE RELISAIT EN MUR. La grille s'ecrivait
   en base trente-six, bornee a trente-cinq « pour que rien ne casse en
   silence ». Elle cassait en silence : PENTE_GAUCHE vaut soixante-
   quatre, sortait « z », se relisait trente-cinq. Toute colline tournee
   vers la gauche se rouvrait fausse. Deux causes : le format ecrivait
   les matieres lui-meme au lieu d'appeler la fonction qui sait les
   ecrire, et il bornait au lieu de refuser. L'alphabet passe a
   soixante-deux caracteres, une case tient toujours en UN, et les
   valeurs 0 a 31 ne bougent pas — un fichier d'avant se relit sans
   migration.

2. NEUF DRAPEAUX DANS UN TABLEAU D'OCTETS. « Moitie haute » vaut deux
   cent cinquante-six et y repassait a zero : une demi-pente haute
   devenait une demi-pente basse, sans erreur, sans avertissement. La
   grille passe en seize bits.

3. LES SIX PORTAGES COMPARAIENT AU CARACTERE « 1 ». Un mur herisse de
   pointes vaut cinq : il etait traversable dans le jeu porte, et solide
   nulle part ailleurs. Ils decodent maintenant les drapeaux et portent
   la regle des pentes. Un banc leur demande, pour chaque matiere et
   chaque colonne de pixels, ou se trouve le sol — sept cent quatre
   reponses, toutes justes.

Et un quatrieme, trouve en mesurant : ON DECOLLAIT POUR UN PIXEL. Une
pente ne descend jamais plus bas que tuile - 1 ; au pied d'une cote il
reste un pixel avant la case plate, et le corps le franchissait en
tombant. Trois images en l'air ne se voient pas, mais elles declenchent
le coyote, passent l'animation en « chute » et rendent possible un saut
aerien juste apres une cote. Le controleur colle au sol jusqu'a la
hauteur qu'il sait monter, et pas un pixel de plus : au bord d'une
falaise il tombe normalement. Les deux cotes sont verifies.

La collision et les tuiles gardent chacune leur copie de la regle — la
collision ne doit rien devoir a l'editeur — mais un banc compare les
deux pour six formes et quatre tailles de tuile.

Version 8 du format. Bancs : 82, 68, 224, 87, 33, 112, 9, fumee 77.
Grille : 61/61, 692 verifications.

+++ src/editeur/edition.ts
@@ -98,7 +98,7 @@ export class Edition {
   private photo: {
     cases: Int32Array[]
     presence: (Uint8Array | null)[]
-    solides: Uint8Array
+    solides: Uint16Array
   } | null = null
   private jeu: Jeu
   private carte: Carte
